Update README.md
This assignment aims to help you learn to:

read a relational schema and analyze instances of the schema

read and apply integrity constraints

express queries and integrity constraints of your own

think about the limits of what can be expressed in relational algebra

Your assignment must be typed to produce a PDF document a1.pdf (hand-written submissions are not acceptable). You may work on the assignment in groups of 1 or 2, and submit a single assignment for the entire group on MarkUs. Partners do not have to be in the same section, but must be in CSC343 at St. George this semester. You must establish your group well before the due date by submitting an incomplete, or even empty, submission. Teaching staﬀ are likely to get quite grumpy if you try to form, or dissolve, a group close to the due date!